license-plate-removal/api-service/requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python-headless
tritonclient[grpc]
python-multipart